+++ leave-service/src/main/java/ddd/leave/interfaces/dto/LeaveDTO.java
@@ -15,7 +15,7 @@ public class LeaveDTO {
     List<ApprovalInfoDTO> historyApprovalInfoDTOList;
     String startTime;
     String endTime;
-    long duration;
+    int duration;
     String status;
 
 }
